requirements.txt
alembic
bcrypt
blinker
click
colorama
Flask
Flask-Bcrypt
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
greenlet
itsdangerous
Jinja2
Mako
MarkupSafe
pyotp
pypng
python-decouple
qrcode
SQLAlchemy
typing_extensions
Werkzeug
WTForms
